+++ frontend/components/NavLink/index.tsx
@@ -1,14 +1,1 @@
-import React from "react";
-import Link from "next/link";
-import { useRouter } from "next/router";
-
-export default ({ href, children }) => {
-  const router = useRouter();
-
-  let className = children.props.className || "";
-  if (router.pathname === href) {
-    className = `${className} active`;
-  }
-
-  return <Link href={href}>{React.cloneElement(children, { className })}</Link>;
-};
+export { default } from "./NavLink";
